Put files into correct folder
